Basic Customers, refactor and fixes
0.5.1
