+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tavern"                                  # Done
-version = "0.4.5"
+version = "0.5.1"
 publish = false
 authors = ["Facundo Venturi"]                    # Done
 edition = "2021"
