+++ mobile/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mobile"
-version = "0.4.5"
+version = "0.5.1"
 edition = "2021"
 publish = false
