+++ mobile/manifest.yaml
@@ -6,6 +6,6 @@ android:
   icon: "ios-src/Assets.xcassets/AppIcon.appiconset/icon_1024x1024.png"
   manifest:
     package: "me.fakuv.tavern" # Done
-    version_code: 9 # Done you should start at 1
+    version_code: 10 # Done you should start at 1
     application:
       label: "Tavern" # Done
