+++ src/ingame/bar.rs
@@ -2,15 +2,32 @@ use bevy::prelude::*;
 
 use crate::{loading::TextureAssets, GameState};
 
-use super::{customer::CustomerBundle, Interactible, InteractibleAction, OnIngameScreen};
+use super::{
+    customer::{generate_customer, CustomerBundle},
+    IngameState, Interactible, InteractibleAction, OnIngameScreen,
+};
 
 // The bar counter
 pub struct BarPlugin;
 
-// Constants
-const BAR_SLOT_LEFT: Vec3 = Vec3::new(-300., -600., 0.);
-const BAR_SLOT_MIDDLE: Vec3 = Vec3::new(0., -600., 0.);
-const BAR_SLOT_RIGHT: Vec3 = Vec3::new(300., -600., 0.);
+// Constants // y: -850.
+const BAR_CUSTOMER_SLOT_LEFT: Vec3 = Vec3::new(-700., -850., 3.);
+const BAR_CUSTOMER_SLOT_MIDDLE: Vec3 = Vec3::new(0., -850., 2.);
+const BAR_CUSTOMER_SLOT_RIGHT: Vec3 = Vec3::new(700., -850., 1.);
+
+impl Plugin for BarPlugin {
+    fn build(&self, app: &mut App) {
+        app //
+            .add_systems(OnEnter(GameState::Playing), setup_bar)
+            .add_systems(
+                Update,
+                (
+                    fill_customer_slots.run_if(in_state(IngameState::Running)),
+                    spawn_customers_in_slots.run_if(in_state(IngameState::Running)),
+                ),
+            );
+    }
+}
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Drink {
@@ -22,18 +39,11 @@ pub enum Drink {
 impl Drink {
     pub fn iterator() -> impl Iterator<Item = (Drink, Vec3)> {
         vec![
-            (Drink::Beer, Vec3::new(750., -615., 4.)),
-            (Drink::Wine, Vec3::new(400., -615., 3.)),
-            (Drink::Whiskey, Vec3::new(50., -615., 2.)),
+            (Drink::Beer, Vec3::new(750., -615., 13.)),
+            (Drink::Wine, Vec3::new(400., -615., 12.)),
+            (Drink::Whiskey, Vec3::new(50., -615., 11.)),
         ]
         .into_iter()
-    }
-}
-
-impl Plugin for BarPlugin {
-    fn build(&self, app: &mut App) {
-        app //
-            .add_systems(OnEnter(GameState::Playing), setup_bar);
     }
 }
 
@@ -45,19 +55,9 @@ struct Bar {
 impl Default for Bar {
     fn default() -> Self {
         Self {
-            customer_slots: BarCustomerSlots {
-                left: None,
-                middle: None,
-                right: None,
-            },
+            customer_slots: BarCustomerSlots::default(),
         }
     }
-}
-
-struct BarCustomerSlots {
-    left: Option<CustomerBundle>,
-    middle: Option<CustomerBundle>,
-    right: Option<CustomerBundle>,
 }
 
 fn setup_bar(mut commands: Commands, textures: Res<TextureAssets>) {
@@ -66,7 +66,7 @@ fn setup_bar(mut commands: Commands, textures: Res<TextureAssets>) {
         .spawn(SpriteBundle {
             texture: textures.bar.clone(),
             transform: Transform {
-                translation: Vec3::new(0., -600., 0.), // y: -540.
+                translation: Vec3::new(0., -733., 10.),
                 scale: Vec3::new(1.5, 1.5, 0.0),
                 ..Default::default()
             },
@@ -78,7 +78,7 @@ fn setup_bar(mut commands: Commands, textures: Res<TextureAssets>) {
         })
         .insert(OnIngameScreen);
 
-    // Bar slots
+    // Drink slots
     for (drink, barrel_pos) in Drink::iterator() {
         commands
             .spawn(SpriteBundle {
@@ -94,5 +94,73 @@ fn setup_bar(mut commands: Commands, textures: Res<TextureAssets>) {
                 action: InteractibleAction::Barrel(drink),
             })
             .insert(OnIngameScreen);
+    }
+}
+
+// Customer slot
+// CustomerBundle is a bundle of components that is used to spawn a customer
+// bool is true if the customer is spawned
+
+struct CustomerSlot {
+    customer: Option<CustomerBundle>,
+    spawned: bool,
+}
+
+impl Default for CustomerSlot {
+    fn default() -> Self {
+        Self {
+            customer: None,
+            spawned: false,
+        }
+    }
+}
+
+#[derive(Default)]
+struct BarCustomerSlots {
+    left: CustomerSlot,
+    middle: CustomerSlot,
+    right: CustomerSlot,
+}
+
+fn fill_customer_slots(mut bar_q: Query<&mut Bar>, textures: Res<TextureAssets>) {
+    let mut bar = bar_q.single_mut();
+
+    if bar.customer_slots.left.customer.is_none() {
+        bar.customer_slots.left.customer = Some(generate_customer(
+            Transform::from_translation(BAR_CUSTOMER_SLOT_LEFT) //
+                .with_scale(Vec3::new(1.5, 1.5, 0.0)),
+            textures.customer.clone(),
+        ));
+    }
+    if bar.customer_slots.middle.customer.is_none() {
+        bar.customer_slots.middle.customer = Some(generate_customer(
+            Transform::from_translation(BAR_CUSTOMER_SLOT_MIDDLE) //
+                .with_scale(Vec3::new(1.5, 1.5, 0.0)),
+            textures.customer.clone(),
+        ));
+    }
+    if bar.customer_slots.right.customer.is_none() {
+        bar.customer_slots.right.customer = Some(generate_customer(
+            Transform::from_translation(BAR_CUSTOMER_SLOT_RIGHT) //
+                .with_scale(Vec3::new(1.5, 1.5, 0.0)),
+            textures.customer.clone(),
+        ));
+    }
+}
+
+fn spawn_customers_in_slots(mut commands: Commands, mut bar_q: Query<&mut Bar>) {
+    let mut bar = bar_q.single_mut();
+
+    if bar.customer_slots.left.customer.is_some() && !bar.customer_slots.left.spawned {
+        commands.spawn(bar.customer_slots.left.customer.take().unwrap());
+        bar.customer_slots.left.spawned = true;
+    }
+    if bar.customer_slots.middle.customer.is_some() && !bar.customer_slots.middle.spawned {
+        commands.spawn(bar.customer_slots.middle.customer.take().unwrap());
+        bar.customer_slots.middle.spawned = true;
+    }
+    if bar.customer_slots.right.customer.is_some() && !bar.customer_slots.right.spawned {
+        commands.spawn(bar.customer_slots.right.customer.take().unwrap());
+        bar.customer_slots.right.spawned = true;
     }
 }
