+++ src/menu/settings.rs
@@ -1,9 +1,6 @@
 use bevy::prelude::*;
 
-use crate::{
-    ingame::{MainCameraIngame, PauseButtonAction},
-    ScreenMode, TEXT_COLOR,
-};
+use crate::{ScreenMode, TEXT_COLOR};
 
 use super::{menu_button, ButtonColors, MenuButtonAction, MenuState};
 
@@ -143,171 +140,6 @@ pub fn settings_main_menu_setup(mut commands: Commands, screen_mode: Res<ScreenM
                 child_builder,
                 "Back",
                 MenuButtonAction::BackToMainMenu,
-                &Style {
-                    // width: Val::Px(300.0),
-                    // height: Val::Px(50.0),
-                    justify_content: JustifyContent::Center,
-                    align_items: AlignItems::Center,
-                    margin: UiRect::top(Val::Px(50.)),
-                    padding: UiRect::axes(Val::Px(15.), Val::Px(10.)),
-                    ..Default::default()
-                },
-                &ButtonColors::default(),
-                &button_text_style,
-            );
-        });
-}
-
-pub fn settings_pause_setup(
-    mut commands: Commands,
-    screen_mode: Res<ScreenMode>,
-    camera_q: Query<&Transform, (With<Camera>, With<MainCameraIngame>)>,
-) {
-    // Transparent Pause background
-    let camera_transform = camera_q.single();
-    commands
-        .spawn(SpriteBundle {
-            transform: Transform {
-                translation: camera_transform.translation.xy().extend(111.),
-                scale: Vec3::new(1920., 1080., 0.0),
-                ..Default::default()
-            },
-            sprite: Sprite {
-                color: Color::Rgba {
-                    red: 0.,
-                    green: 0.,
-                    blue: 0.,
-                    alpha: 0.98,
-                },
-                ..Default::default()
-            },
-            ..Default::default()
-        })
-        .insert(OnSettingsMenuScreen);
-
-    commands
-        .spawn((
-            NodeBundle {
-                style: Style {
-                    width: Val::Percent(100.0),
-                    height: Val::Percent(25.0),
-                    flex_direction: FlexDirection::Column,
-                    align_items: AlignItems::Center,
-                    justify_content: JustifyContent::FlexEnd,
-                    ..Default::default()
-                },
-                // background_color: BackgroundColor(Color::RED),
-                ..Default::default()
-            },
-            OnSettingsMenuScreen,
-        ))
-        .with_children(|child_builder| {
-            child_builder.spawn(TextBundle::from_section(
-                "Settings",
-                TextStyle {
-                    font_size: 120.,
-                    color: TEXT_COLOR,
-                    ..Default::default()
-                },
-            ));
-        });
-
-    commands
-        .spawn((
-            NodeBundle {
-                style: Style {
-                    width: Val::Percent(100.0),
-                    height: Val::Percent(65.0),
-                    flex_direction: FlexDirection::Column,
-                    align_items: AlignItems::Center,
-                    justify_content: JustifyContent::Center,
-                    align_self: AlignSelf::End,
-                    margin: UiRect::bottom(Val::Vh(10.)),
-                    padding: UiRect::vertical(Val::Vh(10.)),
-                    ..Default::default()
-                },
-                // background_color: BackgroundColor(Color::Rgba {
-                //     red: 0.5,
-                //     green: 0.5,
-                //     blue: 0.5,
-                //     alpha: 0.5,
-                // }),
-                ..Default::default()
-            },
-            OnSettingsMenuScreen,
-        ))
-        .with_children(|child_builder| {
-            let button_style = Style {
-                // width: Val::Px(300.0),
-                // height: Val::Px(50.0),
-                justify_content: JustifyContent::Center,
-                align_items: AlignItems::Center,
-                margin: UiRect::bottom(Val::Px(25.)),
-                padding: UiRect::axes(Val::Px(15.), Val::Px(10.)),
-                ..Default::default()
-            };
-            let button_text_style = TextStyle {
-                font_size: 50.0,
-                color: TEXT_COLOR,
-                ..Default::default()
-            };
-
-            child_builder
-                .spawn(NodeBundle {
-                    style: Style {
-                        align_items: AlignItems::Center,
-                        column_gap: Val::Px(15.),
-                        ..default()
-                    },
-                    ..default()
-                })
-                .with_children(|child_builder| {
-                    // Display a label for the current setting
-                    child_builder.spawn(
-                        TextBundle::from_section("Window Mode:", button_text_style.clone())
-                            .with_text_alignment(TextAlignment::Center)
-                            .with_style(Style {
-                                // width: Val::Px(300.0),
-                                // height: Val::Px(50.0),
-                                justify_content: JustifyContent::Center,
-                                align_items: AlignItems::Center,
-                                margin: UiRect {
-                                    right: Val::Px(35.),
-                                    bottom: Val::Px(25.),
-                                    ..Default::default()
-                                },
-                                // padding: UiRect::axes(Val::Px(15.), Val::Px(10.)),
-                                ..Default::default()
-                            }),
-                    );
-                    // Display a button for each possible value
-                    for screen_mode_setting in
-                        [ScreenMode::BorderlessFullscreen, ScreenMode::Windowed]
-                    {
-                        let mut entity = child_builder.spawn((
-                            ButtonBundle {
-                                style: button_style.clone(),
-                                background_color: NORMAL_BUTTON.into(),
-                                ..default()
-                            },
-                            screen_mode_setting,
-                        ));
-                        entity.with_children(|child_builder| {
-                            child_builder.spawn(TextBundle::from_section(
-                                format!("{screen_mode_setting:?}"),
-                                button_text_style.clone(),
-                            ));
-                        });
-                        if *screen_mode == screen_mode_setting {
-                            entity.insert(SelectedOption);
-                        }
-                    }
-                });
-
-            menu_button(
-                child_builder,
-                "Back",
-                PauseButtonAction::BackToPaused,
                 &Style {
                     // width: Val::Px(300.0),
                     // height: Val::Px(50.0),
